Return and refactor component logic back to the components files
