--- packages/ui/src/components/va-tree-view/hooks/useTreeView.ts
@@ -1,43 +1,0 @@
-import { provide, nextTick, ref } from 'vue'
-import { ITreeCategory, TreeView, TreeViewProps, TreeViewProvide, TreeViewKey } from '../types'
-
-export const useTreeView = (props: TreeViewProps): TreeView => {
-  const categories = ref<ITreeCategory[]>([])
-
-  const collapse = () => {
-    nextTick(() => {
-      categories.value.forEach((child: ITreeCategory) => {
-        child.collapse()
-      })
-    })
-  }
-
-  const expand = () => {
-    nextTick(() => {
-      categories.value.forEach((child: ITreeCategory) => {
-        child.expand()
-      })
-    })
-  }
-
-  const onChildMounted = (category: ITreeCategory) => {
-    categories.value.push(category)
-  }
-
-  const onChildUnmounted = (removableCategory: ITreeCategory) => {
-    categories.value = categories.value.filter((category: ITreeCategory) => category !== removableCategory)
-  }
-
-  const treeView: TreeViewProvide = {
-    color: props.color,
-    onChildMounted,
-    onChildUnmounted,
-  }
-
-  provide(TreeViewKey, treeView)
-
-  return {
-    collapse,
-    expand,
-  }
-}
